requirements.txt
# This file specifies the dependencies of the project.
# Install the required libraries by running `pip install -r requirements.txt`.

python-nmap